Drone traversing 1st row completely, 2nd row incompletely. Then deviating off target. [Implemented & Worked] : a) Plant detection using canny and kmeans integration to the whole system went smooth. [Implemented & Failed]: a) Navigation mechanism relying on past plant positons in the same row is flawed because of inaccurate plant seeding + Unpredictible plant growth
(Todo):
a) Design and implement a new navigation system something that gives more weightage to deteted plant and prior plant location only point us in the direction to look for.
b) Using sonar to pinpoint location needs to be improved. One method that can be tested is using a hil climbing algo to find the trough  of sonar values. This is where the plant is actually present.